refactor(ui): merge /advisor into /rebalance as two sections (#1227)
User-confirmed U5c scope (option A only): Rebalance Advisor (rules.yaml
violations → sell priority) and Rebalancing (risk-parity weights) answer
the same question on two axes, so they are now one page — violations
section first (immediate action), weights section second. /advisor
redirects to /rebalance to preserve bookmarks; the sidebar drops the
Advisor entry (the palette follows automatically via NAV_GROUPS).
AdvisorSection moved verbatim to src/app/rebalance/advisor-section.tsx;
scan→explore absorption was declined and stays as-is.

Tests: advisor page tests → redirect lock + await-rendered section
(RSC gotcha pattern); rebalance page tests route both endpoints; e2e
advisor test now asserts the redirect lands on /rebalance.
