+++ frontend/src/__tests__/pages/advisor.test.tsx
@@ -1,5 +1,15 @@
+/**
+ * #1227 U5c: /advisor 는 /rebalance 리다이렉트, AdvisorSection 은 rebalance 로 이동.
+ * 섹션은 async Server Component 라 await-render (RSC coverage gotcha — jsdom 은
+ * 중첩 Suspense child 를 commit 안 함).
+ */
 import { describe, it, expect, vi, beforeEach, type Mock } from "vitest";
 import { render, screen, act } from "@testing-library/react";
+
+const redirectMock = vi.fn();
+vi.mock("next/navigation", () => ({
+  redirect: (...args: unknown[]) => redirectMock(...args),
+}));
 
 // Mock next/link
 vi.mock("next/link", () => ({
@@ -41,28 +51,31 @@ function setupFetchAPI(overrides: { advisor?: unknown } = {}) {
   });
 }
 
-describe("AdvisorPage", () => {
+async function renderSection() {
+  const { AdvisorSection } = await import("@/app/rebalance/advisor-section");
+  const ui = await AdvisorSection();
+  await act(async () => {
+    render(ui);
+  });
+}
+
+describe("AdvisorPage (redirect)", () => {
+  it("redirects /advisor to /rebalance — 북마크 보존 (#1227)", async () => {
+    redirectMock.mockClear();
+    const { default: AdvisorPage } = await import("@/app/advisor/page");
+    AdvisorPage();
+    expect(redirectMock).toHaveBeenCalledWith("/rebalance");
+  });
+});
+
+describe("AdvisorSection (moved to /rebalance)", () => {
   beforeEach(() => {
     vi.restoreAllMocks();
     setupFetchAPI();
   });
 
-  it("renders page heading and description", async () => {
-    const { default: AdvisorPage } = await import("@/app/advisor/page");
-    await act(async () => {
-      render(<AdvisorPage />);
-    });
-
-    expect(screen.getByText("Rebalance Advisor")).toBeInTheDocument();
-    expect(screen.getByText(/투자 규칙 위반 감지/)).toBeInTheDocument();
-  });
-
   it("renders violation count metrics", async () => {
-    const { default: AdvisorPage } = await import("@/app/advisor/page");
-    await act(async () => {
-      render(<AdvisorPage />);
-    });
-
+    await renderSection();
     expect(screen.getByText("총 위반")).toBeInTheDocument();
     expect(screen.getByText("2건")).toBeInTheDocument();
     expect(screen.getByText("Critical")).toBeInTheDocument();
@@ -73,41 +86,25 @@ describe("AdvisorPage", () => {
   });
 
   it("renders total recovery amount", async () => {
-    const { default: AdvisorPage } = await import("@/app/advisor/page");
-    await act(async () => {
-      render(<AdvisorPage />);
-    });
-
+    await renderSection();
     expect(screen.getByText("총 회수 가능")).toBeInTheDocument();
     expect(screen.getByText("$9,090")).toBeInTheDocument();
   });
 
   it("shows critical warning banner when has_critical is true", async () => {
-    const { default: AdvisorPage } = await import("@/app/advisor/page");
-    await act(async () => {
-      render(<AdvisorPage />);
-    });
-
+    await renderSection();
     expect(screen.getByText(/CRITICAL 위반 1건 — 즉시 조치 필요/)).toBeInTheDocument();
   });
 
   it("renders violation type distribution", async () => {
-    const { default: AdvisorPage } = await import("@/app/advisor/page");
-    await act(async () => {
-      render(<AdvisorPage />);
-    });
-
+    await renderSection();
     expect(screen.getByText("위반 유형별 분포")).toBeInTheDocument();
     expect(screen.getByText("stop_loss: 1건")).toBeInTheDocument();
     expect(screen.getByText("position_concentration: 1건")).toBeInTheDocument();
   });
 
   it("renders advisor table description", async () => {
-    const { default: AdvisorPage } = await import("@/app/advisor/page");
-    await act(async () => {
-      render(<AdvisorPage />);
-    });
-
+    await renderSection();
     expect(screen.getByText(/Rebalance Advisor — 매도 우선순위 순/)).toBeInTheDocument();
   });
 
@@ -122,24 +119,14 @@ describe("AdvisorPage", () => {
         has_critical: false,
       },
     });
-
-    const { default: AdvisorPage } = await import("@/app/advisor/page");
-    await act(async () => {
-      render(<AdvisorPage />);
-    });
-
+    await renderSection();
     expect(screen.getByText("READY")).toBeInTheDocument();
     expect(screen.getByText("모든 투자 규칙 준수 중. 위반 사항 없음.")).toBeInTheDocument();
   });
 
   it("handles API failure gracefully", async () => {
     mockFetchAPI = vi.fn().mockRejectedValue(new Error("API down"));
-
-    const { default: AdvisorPage } = await import("@/app/advisor/page");
-    await act(async () => {
-      render(<AdvisorPage />);
-    });
-
+    await renderSection();
     expect(screen.getByText("API 연결 실패. make api 실행 필요.")).toBeInTheDocument();
   });
 });
